+++ src/main/java/cn/tycoding/admin/service/impl/ArticleServiceImpl.java
@@ -1,10 +1,8 @@
 package cn.tycoding.admin.service.impl;
 
 import cn.tycoding.admin.dto.ArticleArchives;
-import cn.tycoding.admin.dto.PageBean;
 import cn.tycoding.admin.entity.*;
-import cn.tycoding.admin.enums.ResultEnums;
-import cn.tycoding.admin.exception.ResultException;
+import cn.tycoding.admin.exception.GlobalException;
 import cn.tycoding.admin.mapper.ArticleMapper;
 import cn.tycoding.admin.service.*;
 import com.alibaba.fastjson.JSON;
@@ -15,9 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 
 /**
  * @auther TyCoding
@@ -25,7 +21,6 @@ import java.util.List;
  */
 @Service
 @SuppressWarnings("all")
-@Transactional
 public class ArticleServiceImpl implements ArticleService {
 
     @Autowired
@@ -54,61 +49,68 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public PageBean findByPage(Article article, int pageCode, int pageSize) {
-        PageHelper.startPage(pageCode, pageSize);
-        Page<Article> page = articleMapper.findByPage(article);
-        List<Article> articleList = page.getResult();
+    public List<Article> findByPage(Article article) {
+        List<Article> articleList = articleMapper.findByPage(article);
         findInit(articleList);
-        return new PageBean(page.getTotal(), articleList);
+        return articleList;
     }
 
-    private void findInit(List<Article> list){
-        for (Article article : list) {
-            List<Category> categoryList = categoryService.findByArticleId(article.getId());
-            if (categoryList.size() > 0) {
-                article.setCategory(categoryList.get(0).getName());
-            }
-            List<Tags> tagsList = tagsService.findByArticleId(article.getId());
-            List<String> stringList = new ArrayList<>();
-            for (Tags tags : tagsList) {
-                stringList.add(tags.getName());
-            }
-            article.setTags(JSON.toJSONString(tagsList));
+    private void findInit(List<Article> list) {
+        if (!list.isEmpty()) {
+            list.forEach(article -> {
+                List<Category> categoryList = categoryService.findByArticleId(article.getId());
+                if (categoryList.size() > 0) {
+                    article.setCategory(categoryList.get(0).getName());
+                }
+                List<Tags> tagsList = tagsService.findByArticleId(article.getId());
+                List<String> stringList = new ArrayList<>();
+                tagsList.forEach(tags -> {
+                    stringList.add(tags.getName());
+                });
+                article.setTags(JSON.toJSONString(tagsList));
+            });
         }
     }
 
     @Override
-    public PageBean findByPageForSite(Integer pageCode, Integer pageSize) {
+    public Map<String, Object> findByPageForSite(int pageCode, int pageSize) {
         PageHelper.startPage(pageCode, pageSize);
         Page<Article> page = articleMapper.findByPageForSite();
         List<Article> articleList = page.getResult();
         findInit(articleList);
-        return new PageBean(page.getTotal(), articleList);
+        Map<String, Object> map = new HashMap<>();
+        map.put("total", page.getTotal());
+        map.put("data", articleList);
+        return map;
     }
 
     @Override
-    public Article findById(long id) {
-        Article article = articleMapper.findById(id);
-        List<Article> articleList = new ArrayList<>();
-        articleList.add(article);
-        findInit(articleList);
-        return article;
+    public Article findById(Long id) {
+        if (!id.equals(null) && id != 0) {
+            Article article = articleMapper.findById(id);
+            List<Article> articleList = new ArrayList<>();
+            articleList.add(article);
+            findInit(articleList);
+            return article;
+        } else {
+            return new Article();
+        }
     }
 
     @Override
+    @Transactional
     public void save(Article article) {
         try {
             if (article.getState() == "1") {
-                //发布
                 article.setPublishTime(new Date());
             }
             article.setEditTime(new Date());
             articleMapper.save(article);
-            long articleId = articleMapper.getLastId(); //查询最新插入文章的ID
-            updateArticleCategoryTags(article, articleId);
+            article.setId(articleMapper.getLastId());
+            updateArticleCategoryTags(article);
         } catch (Exception e) {
             e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
+            throw new GlobalException(e.getMessage());
         }
     }
 
@@ -118,56 +120,63 @@ public class ArticleServiceImpl implements ArticleService {
      * @param article
      * @param id
      */
-    private void updateArticleCategoryTags(Article article, long id) {
-        if (article.getCategory() != null) {
-            //证明新插入的文章有分类信息，将这个文章分类保存到分类表中
-            categoryService.save(new Category(article.getCategory()));
+    @Transactional
+    private void updateArticleCategoryTags(Article article) {
+        if (article.getId() != 0) {
+            if (article.getCategory() != null) {
+                //证明新插入的文章有分类信息，将这个文章分类保存到分类表中
+                categoryService.save(new Category(article.getCategory()));
 
-            //保存了分类信息再保存分类-文章的关联信息
-            Category category = categoryService.findByName(article.getCategory());
-            articleCategoryService.save(new ArticleCategory(id, category.getId()));
-        }
-        if (article.getTags() != null) {
-            //证明新插入的文章有标签数据，将标签数据保存到标签表中
-            List<String> list = (List) JSONArray.parse(article.getTags()); //前端传来的标签是JSON字符串格式的标签名称
-            if (list.size() > 0) {
-                for (String name : list) {
-                    tagsService.save(new Tags(name));
-                    Tags tags = tagsService.findByName(name); //因为标签是多个的，需要依次将标签信息保存到标签表中
+                //保存了分类信息再保存分类-文章的关联信息
+                Category category = categoryService.findByName(article.getCategory());
+                articleCategoryService.save(new ArticleCategory(article.getId(), category.getId()));
+            }
+            if (article.getTags() != null) {
+                //证明新插入的文章有标签数据，将标签数据保存到标签表中
+                List<String> list = (List) JSONArray.parse(article.getTags()); //前端传来的标签是JSON字符串格式的标签名称
+                if (list.size() > 0) {
+                    list.forEach(name -> {
+                        tagsService.save(new Tags(name));
+                        Tags tags = tagsService.findByName(name); //因为标签是多个的，需要依次将标签信息保存到标签表中
 
-                    if (tags != null) {
-                        //说明该标签插入成功或已存在，建立标签-文章关联信息
-                        articleTagsService.save(new ArticleTags(id, tags.getId()));
-                    }
+                        if (tags != null) {
+                            //说明该标签插入成功或已存在，建立标签-文章关联信息
+                            articleTagsService.save(new ArticleTags(article.getId(), tags.getId()));
+                        }
+                    });
                 }
             }
         }
     }
 
     @Override
+    @Transactional
     public void update(Article article) {
         try {
             articleMapper.update(article);
-            updateArticleCategoryTags(article, article.getId());
+            updateArticleCategoryTags(article);
         } catch (Exception e) {
             e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
+            throw new GlobalException(e.getMessage());
         }
     }
 
     @Override
+    @Transactional
     public void delete(Long... ids) {
-        try {
-            for (long id : ids) {
-                articleMapper.delete(id);
-                //删除文章-分类表的关联
-                articleCategoryService.deleteByArticleId(id);
-                //删除文章-标签表的关联
-                articleTagsService.deleteByArticleId(id);
+        if (ids.length > 0 && !ids.equals(null)) {
+            try {
+                for (long id : ids) {
+                    articleMapper.delete(id);
+                    //删除文章-分类表的关联
+                    articleCategoryService.deleteByArticleId(id);
+                    //删除文章-标签表的关联
+                    articleTagsService.deleteByArticleId(id);
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new GlobalException(e.getMessage());
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
         }
     }
 
@@ -181,35 +190,37 @@ public class ArticleServiceImpl implements ArticleService {
         List<ArticleArchives> articleArchivesList = new ArrayList<ArticleArchives>();
         try {
             List<String> dates = articleMapper.findArchivesDates();
-            for (String date : dates) {
+            dates.forEach(date -> {
                 List<Article> articleList = articleMapper.findArchivesByDate(date);
                 ArticleArchives articleArchives = new ArticleArchives(date, articleList);
                 articleArchivesList.add(articleArchives);
-            }
+            });
         } catch (Exception e) {
             e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
+            throw new GlobalException(e.getMessage());
         }
         return articleArchivesList;
     }
 
     @Override
     public List<Article> findFuzzyByTitle(String title) {
-        try {
+        if (!title.isEmpty()) {
             return articleMapper.findFuzzyByTitle(title);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
+        } else {
+            return new ArrayList<>();
         }
     }
 
     @Override
-    public void addEyeCount(long id) {
-        try {
-            articleMapper.addEyeCount(id);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new ResultException(ResultEnums.INNER_ERROR);
+    @Transactional
+    public void addEyeCount(Long id) {
+        if (!id.equals(null) && id != 0) {
+            try {
+                articleMapper.addEyeCount(id);
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new GlobalException(e.getMessage());
+            }
         }
     }
 }
