+++ src/main/java/cn/tycoding/admin/service/impl/UserServiceImpl.java
@@ -1,14 +1,10 @@
 package cn.tycoding.admin.service.impl;
 
-import cn.tycoding.admin.dto.PageBean;
 import cn.tycoding.admin.dto.PasswordHelper;
 import cn.tycoding.admin.entity.User;
-import cn.tycoding.admin.enums.ResultEnums;
-import cn.tycoding.admin.exception.ResultException;
+import cn.tycoding.admin.exception.GlobalException;
 import cn.tycoding.admin.mapper.UserMapper;
 import cn.tycoding.admin.service.UserService;
-import com.github.pagehelper.Page;
-import com.github.pagehelper.PageHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +17,6 @@ import java.util.List;
  */
 @Service
 @SuppressWarnings("all")
-@Transactional
 public class UserServiceImpl implements UserService {
 
     @Autowired
@@ -41,52 +36,65 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public PageBean findByPage(User user, int pageCode, int pageSize) {
-        PageHelper.startPage(pageCode, pageSize);
-        Page page = userMapper.findByPage(user);
-        return new PageBean(page.getTotal(), page.getResult());
+    @Transactional
+    public List<User> findByPage(User user) {
+        return userMapper.findByPage(user);
     }
 
     @Override
-    public User findById(long id) {
+    public User findById(Long id) {
         return userMapper.findById(id);
     }
 
     @Override
+    @Transactional
     public void save(User user) {
         try {
             passwordHelper.encryptPassword(user); //加密
             userMapper.save(user);
         } catch (Exception e) {
-            throw new ResultException(ResultEnums.INNER_ERROR);
+            e.printStackTrace();
+            throw new GlobalException(e.getMessage());
         }
     }
 
     @Override
+    @Transactional
     public void update(User user) {
-        try {
-            if (user.getPassword() != null && !"".equals(user.getPassword())) {
-                passwordHelper.encryptPassword(user); //加密
+        if (user.getId() != 0) {
+            try {
+                if (user.getPassword() != null && !"".equals(user.getPassword())) {
+                    passwordHelper.encryptPassword(user); //加密
+                }
+                userMapper.update(user);
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new GlobalException(e.getMessage());
             }
-            userMapper.update(user);
-        } catch (Exception e) {
-            throw new ResultException(ResultEnums.INNER_ERROR);
         }
     }
 
     @Override
+    @Transactional
     public void delete(Long... ids) {
-        try {
-            for (long id : ids) {
-                userMapper.delete(id);
+        if (!ids.equals(null) && ids.length > 0) {
+            try {
+                for (long id : ids) {
+                    userMapper.delete(id);
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new GlobalException(e.getMessage());
             }
-        } catch (Exception e) {
-            throw new ResultException(ResultEnums.INNER_ERROR);
         }
     }
 
     @Override
     public User findByName(String username) {
-        return userMapper.findByName(username);
+        if (!username.isEmpty()) {
+            return userMapper.findByName(username);
+        } else {
+            return new User();
+        }
     }
 }
